requirements.txt
numpy
spleeter
pydub
soundfile
pyAudioAnalysis
requests
librosa
moviepy
pytube
python-dotenv
srt